Fix: Remove incorrect analyses updated_at trigger
The analyses table doesn't have an updated_at column, but init.sql
was creating a trigger that tried to update it, causing:
'record "new" has no field "updated_at"' errors.

Changes:
- Dropped update_analyses_updated_at trigger from database
- Commented out trigger creation in init.sql
- Created migration 002_fix_analyses_trigger.sql
- Added note explaining why trigger is disabled

This fixes analysis creation which was failing with P2010 error.

If updated_at is added to analyses in the future, the trigger
can be uncommented and recreated.